Move ahklsConfig object to config.ts
