+++ server/src/commandProvider.ts
@@ -1,11 +1,11 @@
 import { CancellationToken, ExecuteCommandParams, Position, Range, SymbolKind } from 'vscode-languageserver';
 import {
 	AhkSymbol, ClassNode, FuncNode, Lexer, Property, Token, Variable,
-	connection, ahklsConfig, find_class, generate_type_annotation,
+	connection, find_class, generate_type_annotation,
 	join_types, lexers, parse_include, restorePath, semanticTokensOnFull,
 	traverse_include, update_include_cache
 } from './common';
-import { CfgKey, getCfg } from '../../util/src/config';
+import { CfgKey, getCfg, ahklsConfig } from '../../util/src/config';
 
 function checkCommand(cmd: string) {
 	if (getCfg(ahklsConfig, CfgKey.Commands)?.includes(cmd))
